Make it clear CustomQuerier is just for test code
