+++ contracts/reflect/src/lib.rs
@@ -1,7 +1,9 @@
 pub mod contract;
 pub mod msg;
-pub mod query;
 pub mod state;
+
+#[cfg(not(target_arch = "wasm32"))]
+pub mod testing;
 
 #[cfg(target_arch = "wasm32")]
 mod wasm {
