+++ contracts/reflect/src/msg.rs
@@ -1,7 +1,7 @@
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
-use cosmwasm_std::{Binary, CosmosMsg, HumanAddr};
+use cosmwasm_std::{Binary, CosmosMsg, HumanAddr, QueryRequest};
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct InitMsg {}
@@ -39,4 +39,26 @@ impl Into<CosmosMsg<CustomMsg>> for CustomMsg {
     fn into(self) -> CosmosMsg<CustomMsg> {
         CosmosMsg::Custom(self)
     }
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+#[serde(rename_all = "snake_case")]
+/// CustomQuery is an override of QueryRequest::Custom to show this works and can be extended in the contract
+pub enum CustomQuery {
+    Ping {},
+    Capital { text: String },
+}
+
+// TODO: do we want to standardize this somehow for all?
+impl Into<QueryRequest<CustomQuery>> for CustomQuery {
+    fn into(self) -> QueryRequest<CustomQuery> {
+        QueryRequest::Custom(self)
+    }
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+#[serde(rename_all = "snake_case")]
+// All return values of CustomQuery are CustomResponse
+pub struct CustomResponse {
+    pub msg: String,
 }
